new build quiz view

+++ quiz/models.py
@@ -272,7 +272,7 @@ class Question(models.Model):
     idealDuration = models.DurationField(default=timedelta(seconds=120), blank=True)
 
     tags = models.ManyToManyField(Tag, related_name='tags', blank=True)
-    level = models.IntegerField(default=2, blank=True)
+    level = models.FloatField(default=2.0, blank=True)
     hint = models.TextField(null=True, blank=True)
 
     def __str__(self):
